relocate files and change names
