+++ composeApp/src/androidMain/AndroidManifest.xml
@@ -31,7 +31,7 @@
             </intent-filter>
         </activity>
         <service
-            android:name=".components.MyHostApduService"
+            android:name=".services.HostApduService"
             android:exported="true"
             android:permission="android.permission.BIND_NFC_SERVICE">
             <intent-filter>
