--- composeApp/src/nativeMain/kotlin/com/github/familyvault/components/NFCManager.native.kt
@@ -1,4 +1,0 @@
-package com.github.familyvault.components
-
-actual fun unregisterApp() {
-}
